Move codec comparison to Capabilities.js
